+++ FE/src/components/planner/day/TodoList.tsx
@@ -1,16 +1,8 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import styles from "@styles/planner/day.module.scss";
-import Text from "@components/common/Text";
-import { AddCircleOutline, DeleteOutlined } from "@mui/icons-material";
-import { todoListType, categoryType } from "@util/planner.interface";
-import { todoData_list as data, todoData_category } from "@util/data/DayTodos";
-
-interface TodoItemProps {
-  item: todoListType;
-  insertTodo: () => void;
-  updateTodo: (todo: todoListType) => void;
-  deleteTodo: (id: number) => void;
-}
+import TodoItem from "@components/planner/day/TodoItem";
+import { todoListType } from "@util/planner.interface";
+import { todoData_list as data } from "@util/data/DayTodos";
 
 const TodoDataDefalut: todoListType = {
   todoId: -1,
@@ -19,71 +11,6 @@ const TodoDataDefalut: todoListType = {
   todoContent: "",
   todoStatus: 0,
   isPossible: false,
-};
-
-const TodoItem = ({ item, insertTodo, updateTodo, deleteTodo }: TodoItemProps) => {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const { todoId, categoryName, todoContent, todoStatus, isPossible } = item;
-  const checkedArr = [" ", "O", "X"];
-  const [text, setText] = useState(todoContent);
-  const modifyText = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setText(e.target.value);
-  };
-
-  const handleOnBlur = () => {
-    saveTodo();
-  };
-
-  const handleOnKeyPress = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Enter") {
-      saveTodo();
-    }
-  };
-
-  const addTodo = (possible: boolean | undefined) => {
-    if (possible != undefined && !possible) insertTodo();
-  };
-
-  const removeTodo = () => {
-    deleteTodo(todoId);
-    setText("");
-  };
-
-  const saveTodo = () => {
-    updateTodo({ ...item, todoContent: text });
-  };
-
-  return (
-    <div
-      className={`${styles["todo-item"]} ${!isPossible && styles["todo-item--disable"]}`}
-      onClick={() => addTodo(isPossible)}
-    >
-      <div className={styles["todo-item__category"]}>
-        {isPossible ? <Text>{categoryName}</Text> : <AddCircleOutline />}
-      </div>
-      <div className={styles["todo-item__context"]}>
-        {isPossible ? (
-          <div className={styles["todo-item__context__possiable"]}>
-            <input
-              ref={inputRef}
-              value={text}
-              onChange={modifyText}
-              onBlur={handleOnBlur}
-              onKeyDown={handleOnKeyPress}
-            />
-            <div onClick={removeTodo}>
-              <DeleteOutlined />
-            </div>
-          </div>
-        ) : (
-          <AddCircleOutline />
-        )}
-      </div>
-      <div className={styles["todo-item__checked"]}>
-        <Text types="semi-medium"> {checkedArr[todoStatus]}</Text>
-      </div>
-    </div>
-  );
 };
 
 const TodoList = () => {
